fix(docs): drop discount names when the snapshot no longer agrees with the order
InvoiceSnapshot::discountNote() named every order_discounts row whenever the
live discount total was positive, even when only SOME discounted lines were
edited away. OrderDiscount carries no per-item linkage, so a partial edit can
leave the live total below the sum of still-unedited snapshot rows with no
way to tell which source is still true — naming one anyway is a materially
misleading claim on a tax document. The note now compares the live total to
the snapshot sum and only names discounts when they agree exactly; otherwise
it prints a generic, code-free clause with the correct (live) amount.

Also widens documents.discount_total to unsignedBigInteger, matching the
width documents.total already uses on this table.

+++ Modules/Docs/Database/Migrations/2026_07_28_130000_add_discount_note_to_documents.php
@@ -18,13 +18,19 @@ use Illuminate\Support\Facades\Schema;
  * proforma snapshot that never sets them simply gets the column defaults
  * (0 / null), so InvoiceIssuer is the only builder that has to know about
  * this wave's addition.
+ *
+ * `discount_total` is `unsignedBigInteger`, matching the width `total` on
+ * this same table was widened to by the 2026-07-22 migration — unsigned
+ * because, unlike `total`, nothing ever writes a negative discount_total
+ * (a credit note never sets it), but the same 8-byte width as its sibling
+ * money column rather than the narrower 4-byte default.
  */
 return new class extends Migration
 {
     public function up(): void
     {
         Schema::table('documents', function (Blueprint $table) {
-            $table->unsignedInteger('discount_total')->default(0)->after('total');
+            $table->unsignedBigInteger('discount_total')->default(0)->after('total');
             $table->string('discount_note', 500)->nullable()->after('discount_total');
         });
     }
